backend/pyproject.toml
[project]
name = "backend"
version = "0.0.1"
requires-python = ">=3.14,<3.15"
dependencies = [
    "uvicorn",
    "django-ninja",
    "python-dotenv",
    "django-simpleui",
    "django-ninja-extra",
    "psycopg[binary,pool]",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
